ProviderIcon.tsx
